cambio de tsx a jsx
