--- src/models/restaurant.ts
@@ -1,7 +1,0 @@
-export interface Restaurant {
-    id?: number;
-    name?: string;
-    address:string;
-    phone?: string;
-    email?: string;
-}
